Gui_Build_V1_0704/requirements.txt
streamlit>=1.24.0
tensorflow>=2.16.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.3